refactor: Update arweaveAliasDispatcher to use arweaveGetAlias in arweave_alias.go
This commit refactors the arweaveAliasDispatcher function in arweave_alias.go to use the arweaveGetAlias function instead. The previous implementation was using the joinAlertGroup function, which was incorrect. This change ensures that the correct function is called and improves the accuracy of the code.

Refactoring the code in this way enhances the functionality and maintainability of the arweave alias feature in the Telegram service.