fix: Include link to all alias records in arweave_alias.go
This commit adds a link to the list of all alias records in the arweave_alias.go file. The link points to the URL "https://ario.laisky.com/dns" and is included in the help message for the arweave alias feature in the Telegram service.

This change provides users with easy access to view all alias records, enhancing the functionality and usability of the feature.